+++ administrator/components/com_j2xml/src/Model/SendModel.php
@@ -18,11 +18,6 @@
 // no direct access
 namespace Joomla\Component\J2xml\Administrator\Model;
 
-use Joomla\CMS\Component\ComponentHelper;
-use Joomla\CMS\Log\Log;
-use Joomla\CMS\Log\LogEntry;
-use Joomla\CMS\MVC\Model\FormModel;
-
 \defined('_JEXEC') or die;
 
 /**
@@ -30,133 +25,7 @@ use Joomla\CMS\MVC\Model\FormModel;
  *
  * @since 3.9.0
  */
-class SendModel extends FormModel
+class SendModel extends AbstractFormModel
 {
-
-    /**
-     * The model context
-     *
-     * @var string
-     */
-    protected $context = 'j2xml';
-
-    /**
-     * Constructor.
-     *
-     * @param array $config
-     *          An optional associative array of configuration settings.
-     *
-     * @see \Joomla\CMS\MVC\Model\BaseModel
-     * @since 3.9.0
-     */
-    public function __construct($config = [])
-    {
-        Log::add(new LogEntry(__METHOD__, Log::DEBUG, 'com_j2xml'));
-
-        $layout = \Joomla\CMS\Factory::getApplication()->getInput()->get('layout', 'default');
-        if ($layout != 'default')
-        {
-            $this->context .= '.' . $layout;
-        }
-
-        parent::__construct($config);
-    }
-
-    /**
-     * Method to get the record form.
-     *
-     * @param array $data
-     *          Data for the form.
-     * @param boolean $loadData
-     *          True if the form is to load its own data (default case), false if not.
-     *
-     * @return JForm|boolean A JForm object on success, false on failure
-     *
-     * @since 3.9.0
-     */
-    public function getForm($data = [], $loadData = true)
-    {
-        Log::add(new LogEntry(__METHOD__, Log::DEBUG, 'com_j2xml'));
-
-        try
-        {
-            $form = $this->loadForm($this->context, 'send', [
-                'control' => 'jform',
-                'load_data' => false
-            ]);
-
-            $layout = \Joomla\CMS\Factory::getApplication()->getInput()->get('layout', 'default');
-            if ($layout != 'default')
-            {
-                $form->loadFile('send_' . $layout);
-
-                if ($layout != 'users')
-                {
-                    $form->loadFile('send_users');
-                    /* if ($layout == 'contact')
-                     {
-                     $form->setFieldAttribute('send_contacts', 'type', 'hidden');
-                     } */
-                }
-                /* else
-                 {
-                 $form->setFieldAttribute('send_users', 'type', 'hidden');
-                 } */
-            }
-
-            if ($loadData)
-            {
-                // Get the data for the form.
-                $data = $this->loadFormData();
-            }
-            else
-            {
-                $data = [];
-            }
-
-            // Allow for additional modification of the form, and events to be triggered.
-            // We pass the data because plugins may require it.
-            $this->preprocessForm($form, $data);
-
-            // Load the data into the form after the plugins have operated.
-            $form->bind($data);
-        }
-        catch (\Exception $e)
-        {
-            $this->setError($e->getMessage());
-
-            return false;
-        }
-
-        return $form;
-    }
-
-    /**
-     * Method to get the data that should be injected in the form.
-     *
-     * @return mixed The data for the form.
-     *
-     * @since 3.9.0
-     */
-    protected function loadFormData()
-    {
-        Log::add(new LogEntry(__METHOD__, Log::DEBUG, 'com_j2xml'));
-
-        // Check the session for previously entered form data.
-        $data   = \Joomla\CMS\Factory::getApplication()->getUserState('com_j2xml.send.data', []);
-        Log::add(new LogEntry('getUserState(\'com_j2xml.send.data\'): ' . print_r($data, true), Log::DEBUG, 'com_j2xml'));
-        $jform  = [];
-        foreach($data as $k => $v)
-        {
-            $jform['send_' . $k] = $v;
-        }
-
-        $params = ComponentHelper::getParams('com_j2xml');
-        $data   = array_merge($params->toArray(), $jform);
-        Log::add(new LogEntry('data: ' . print_r($data, true), Log::DEBUG, 'com_j2xml'));
-
-        $this->preprocessData($this->context, $data);
-
-        return $data;
-    }
+    protected string $formType = 'send';
 }